+++ Xplat/src/main/java/vazkii/botania/data/BlockTagProvider.java
@@ -291,6 +291,9 @@ public class BlockTagProvider extends IntrinsicHolderTagsProvider<Block> {
 		tag(BotaniaTags.Blocks.AGRICARNATION_GROWTH_CANDIDATE).addTag(BotaniaTags.Blocks.AGRICARNATION_APPLY_BONEMEAL);
 		tag(BotaniaTags.Blocks.AGRICARNATION_GROWTH_EXCLUDED).add(Blocks.RED_MUSHROOM, Blocks.BROWN_MUSHROOM, Blocks.MANGROVE_LEAVES);
 
+		// for Diagonal Windows compatibility (DW adds its version of the block to the tag automatically)
+		tag(BotaniaTags.Blocks.FEL_BLAZE_BASE).add(Blocks.IRON_BARS);
+
 		registerMiningTags();
 	}