+++ web/changelog.md
@@ -26,6 +26,7 @@ In the meantime, Botania for Minecraft 1.20.1 may still receive updates for bug 
 * Fix: The information whether an item entity was created via mana infusion or via the runic altar was deserialized incorrectly
 * Fix: Mobs were able to spawn on spectral platforms
 * Fix: Floating special flowers now show the type and reference lines, like their non-floating versions do
+* Fix: Diagonal Windows iron bars now also work for summoning fel blazes
 
 ---